requirements_agentic.txt
# Updated requirements.txt for True Agentic System

# Core Framework
flask==3.0.3
werkzeug==3.0.4

# AI/ML
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langgraph==0.2.28

# Document Processing
pillow==10.4.0
pypdf==4.3.1
python-docx==1.1.2

# Utilities
requests==2.32.3
python-dotenv==1.0.1
typing-extensions==4.12.2
flask-cors==4.0.0

# Database (Supabase PostgreSQL with pgvector)
psycopg2-binary==2.9.9

# Embedding Providers (Alternative to OpenAI)
sentence-transformers==2.2.2  # Free, local embeddings (HuggingFace)
torch==2.1.0  # Required by sentence-transformers

# Logging and Monitoring (already included in standard library, but listed for clarity)
# logging - standard library
# threading - standard library
# queue - standard library

# Agentic System Dependencies
# (Most are already included above)

# Optional: For production agentic systems
# redis==5.0.8  # For distributed memory
# chromadb==0.4.22  # For vector-based experience database
# celery==5.4.0  # For async task processing
# prometheus-client==0.21.0  # For metrics

# Development
# pytest==8.3.3
# pytest-cov==5.0.0
# black==24.8.0
# ruff==0.6.8
